rename tiptap editor to CoachingNotes
